event triggered rightly

+++ gamification-impl/src/main/java/ch/heig/amt/gamification/repositories/UserRepository.java
@@ -3,7 +3,7 @@ package ch.heig.amt.gamification.repositories;
 import ch.heig.amt.gamification.entities.UserEntity;
 import org.springframework.data.repository.CrudRepository;
 
-public interface UserRepository extends CrudRepository<UserEntity, Long> {
+public interface UserRepository extends CrudRepository<UserEntity, Integer> {
     Iterable<UserEntity> findAllByApplicationEntity_ApiKey(String apiKey);
     UserEntity findByUsernameAndApplicationEntity_ApiKey(String name, String ApiKey);
 }
